+++ desktop/cmp/error/ErrorMessage.js
@@ -19,7 +19,16 @@ import './ErrorMessage.scss';
  */
 export const [ErrorMessage, errorMessage] = hoistCmp.withFactory({
     className: 'xh-error-message',
-    render({className, error, message, title, actionFn, actionButtonProps}, ref) {
+    render({
+        className,
+        model,
+        error = model?.error,
+        message,
+        title,
+        actionFn,
+        actionButtonProps
+    }, ref) {
+
         // TODO - remove in v42+
         apiDeprecated(actionFn, 'actionFn', "Use 'actionButtonProps' instead");
         if (actionFn) {
@@ -60,7 +69,8 @@ ErrorMessage.propTypes = {
     actionButtonProps: PT.object,
 
     /**
-     *  Error to display. If null or undefined this component will not be displayed.
+     *  Error to display. If undefined, this component will look for an error property on its model.
+     *  If no error is found, this component will not be displayed.
      */
     error: PT.oneOfType([PT.instanceOf(Error), PT.object,  PT.string]),